Separa os dois middlewares em dois arquivos
